Use FastAPI for viewer
