+++ requirements.txt
@@ -6,6 +6,7 @@ click>=8.1.3,!=8.2.0
 debugpy
 docstring-parser>=0.16
 exceptiongroup>=1.0.2; python_version < '3.11'
+fastapi==0.116.1
 fsspec>=2023.1.0,<=2025.3.0 # align with hf datasets to prevent pip errors
 httpx
 ijson>=3.2.0
@@ -30,4 +31,5 @@ sniffio
 tenacity
 textual>=0.86.2,<v3.0.0 # https://github.com/UKGovernmentBEIS/inspect_ai/issues/1891
 typing_extensions>=4.9.0
+uvicorn>=0.30.0
 zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
